# swagger.yaml
swagger: "2.0"
info:
  title: "API Gêmeos Brasil"
  description: |
    Sistema de Rifas - API REST completa para integração front-end.

    ## Autenticação
    Endpoints protegidos exigem JWT no header:
    ```
    Authorization: Bearer {token}
    ```
    O token é obtido via `/api/auth/login` ou `/api/auth/registro`.
  version: "2.1"
  contact:
    email: "contato@gemeosbrasil.com"
host: "localhost:5000"
basePath: "/"
schemes:
  - "http"
  - "https"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
    description: "JWT Bearer Token. Formato: Bearer {token}"
...
paths:
...
  /api/painel-secreto-x9/usuarios:
    get:
      tags: ["Admin"]
      summary: "Listar todos os usuários"
      description: "Prefixo configurável via variável ADMIN_ROUTE_SECRET."
      security:
        - Bearer: []
      parameters:
        - name: "page"
          in: "query"
          type: "integer"
          default: 1
        - name: "per_page"
          in: "query"
          type: "integer"
          default: 20
      responses:
        200:
          description: "Lista de usuários"
          schema:
            type: "object"
            properties:
              usuarios:
                type: "array"
                items:
                  $ref: "#/definitions/Usuario"
              total:
                type: "integer"
              paginas:
                type: "integer"
              pagina_atual:
                type: "integer"
        403:
          description: "Acesso negado"
  /api/painel-secreto-x9/dashboard:
    get:
      tags: ["Admin"]
      summary: "Dados do dashboard administrativo"
      security:
        - Bearer: []
      responses:
        200:
          description: "Estatísticas do sistema"
          schema:
            type: "object"
            properties:
              stats:
                type: "object"
                properties:
                  total_usuarios:
                    type: "integer"
                  total_campanhas:
                    type: "integer"
                  campanhas_ativas:
                    type: "integer"
                  receita_total:
                    type: "number"
              ultimas_vendas:
                type: "array"
                items:
                  type: "object"
                  properties:
                    id:
                      type: "string"
                    campanha:
                      type: "string"
                    usuario:
                      type: "string"
                    valor:
                      type: "number"
                    data:
                      type: "string"
        403:
          description: "Acesso negado"
  /api/painel-secreto-x9/campanhas/{campanha_id}/auditoria:
    get:
      tags: ["Admin"]
      summary: "Auditoria de sorteio da campanha"
      description: |
        Retorna o histórico de `AdminLog` filtrado pela campanha.
        Inclui todas as ações registradas (criação, atualização, definição de ganhador, etc.).
        Prefixo do admin configurável via variável `ADMIN_ROUTE_SECRET`.
      security:
        - Bearer: []
      parameters:
        - name: "campanha_id"
          in: "path"
          required: true
          type: "string"
          description: "public_id (UUID) ou id interno da campanha"
        - name: "page"
          in: "query"
          type: "integer"
          default: 1
        - name: "per_page"
          in: "query"
          type: "integer"
          default: 20
          description: "Máximo: 50"
      responses:
        200:
          description: "Histórico de auditoria da campanha"
          schema:
            type: "object"
            properties:
              campanha_id:
                type: "string"
                format: "uuid"
              campanha_titulo:
                type: "string"
              auditoria:
                type: "array"
                items:
                  $ref: "#/definitions/AdminLog"
              total:
                type: "integer"
              pagina:
                type: "integer"
              por_pagina:
                type: "integer"
              paginas:
                type: "integer"
        401:
          description: "Não autenticado"
        403:
          description: "Acesso negado"
        404:
          description: "Campanha não encontrada"

  # ==========================================================================
  # ARTIGOS
  # ==========================================================================
...
definitions:
...
  Usuario:
    type: "object"
    properties:
      id:
        type: "integer"
      nome:
        type: "string"
      telefone:
        type: "string"
      email:
        type: "string"
        description: "Mascarado: jo***@email.com"
      cpf:
        type: "string"
        description: "Mascarado: 123.***.***-**"
      cidade:
        type: "string"
      estado:
        type: "string"
      is_admin:
        type: "boolean"
      criado_em:
        type: "string"
        format: "date-time"
...
  AdminLog:
    type: "object"
    description: "Entrada de log de ação administrativa"
    properties:
      id:
        type: "integer"
      admin_name:
        type: "string"
        description: "Nome do administrador que executou a ação"
      action:
        type: "string"
        example: "Definição de Ganhador"
      details:
        type: "string"
        description: "Detalhes adicionais em texto livre"
      ip_address:
        type: "string"
        example: "192.168.0.1"
      created_at:
        type: "string"
        format: "date-time"